% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_ww.R
\name{adam_ww}
\alias{adam_ww}
\title{ADaM PASI - Wonderful Wednesdays}
\source{
Derived from
  Wonderful Wednesdays GitHub data repository (2021-04-14).
}
\usage{
adam_ww(n = NULL, seed = 1)
}
\arguments{
\item{n}{(`numeric`)\cr number of patients to keeps.}

\item{seed}{a single value, interpreted as an integer, or \code{NULL}
    (see \sQuote{Details}).}
}
\value{
A list with two ADaM datasets:
\describe{
  \item{ADSL}{subject level dataset}
  \item{ADPASI}{PASI records, follows basic data structure (BDS)}
}
}
\description{
The Psoriasis Area and Severity Index (PASI) is a continuous measurement
of severity, the lower, the better the patient condition. The
derived dataset represents patient level follow-up during 8 visits over a
year of monitoring.
}
\examples{

adam_ww(2)

}
