% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing.R
\name{add_missing}
\alias{add_missing}
\title{Factor: NA is "Missing" level}
\usage{
add_missing(x, missing_lvl = "Missing")
}
\arguments{
\item{x}{(`factor`)}

\item{missing_lvl}{(`character`)}
}
\description{
Assign the value "Missing" to missing values of a factor. Built as [addNA()].
}
\examples{

animals <- as.factor(c("cat", "dog", NA))
animals
add_missing(animals)

}
