% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_slate.R
\name{slate}
\alias{slate}
\alias{clean_slate}
\alias{add_figure}
\alias{add_header}
\alias{add_footer}
\alias{add_title}
\alias{add_note}
\title{Annotation Slate}
\usage{
clean_slate(
  margin = grid::unit(c(t = 2.54, r = 4.2, b = 2.54, l = 2.54), units = "cm"),
  gp = grid::gpar(lineheight = 1, fontfamily = "Courier", fontsize = 9)
)

add_figure(page, gr, guide = FALSE, width = 1, height = 1)

add_header(page, left, right, ...)

add_footer(page, left, right, ...)

add_title(page, text, guide = FALSE)

add_note(
  page,
  text,
  guide = FALSE,
  str_width = 80,
  x = 0.1,
  skipped_lines = c(1, 0)
)
}
\arguments{
\item{margin}{(`unit`)\cr top, right, bottom, left margins.}

\item{gp}{(`list`)\cr graphical parameter, see [grid::gpar()].}

\item{page}{(`grob`)\cr a page as initialised by `clean_slate()`.}

\item{gr}{(`grob`)\cr a graphical object.}

\item{guide}{(`logical`)\cr display rectangle delimitating a page area.}

\item{width, height}{(`numeric` or `unit`).}

\item{left, right, text}{(`character`)\cr annotation texts.}

\item{...}{arguments passed to other functions.}

\item{str_width}{(`numeric`)\cr maximal width.}

\item{x}{(`numeric` or `unit`)\cr horizontal position.}

\item{skipped_lines}{(`numerix`)\cr number of lines to skip.}
}
\description{
Annotation Slate
}
\section{Functions}{
\itemize{
\item \code{clean_slate}: start an annotation page.

\item \code{add_figure}: add a figure to a clean page.

\item \code{add_header}: add a header.

\item \code{add_footer}: add a footer.

\item \code{add_title}: add a title.

\item \code{add_note}: add a note.
}}

\examples{

library(magrittr)
library(ggplot2)

clean_slate() \%>\%
  add_figure(gr = circleGrob()) \%>\%
  grid.draw()

# Default margins are for an A4 page, change it for the plot viewer
clean_slate(margin = unit(rep(0.05, 4), units = "npc")) \%>\%
  add_figure(gr = circleGrob()) \%>\%
  grid.draw()

# Of course, it works with ggplot objects
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
gg <- ggplot(dsamp, aes(carat, price)) +
  geom_point(aes(colour = clarity), show.legend = FALSE)

clean_slate(margin = unit(rep(0.05, 4), units = "npc")) \%>\%
  add_figure(gr = gg) \%>\%
  grid.draw()

# Then add annotation elements, for instance, a title ...
clean_slate(margin = unit(rep(0.05, 4), units = "npc")) \%>\%
  add_figure(gr = gg, height = .5) \%>\%
  add_title("Figure 1") \%>\%
  grid.draw()

# ... and / or other elements ...
clean_slate(margin = unit(rep(0.05, 4), units = "npc")) \%>\%
  add_figure(gr = gg) \%>\%
  add_header(left = "Topic", right = "CONFIDENTIAL") \%>\%
  add_title("Figure 1") \%>\%
  add_note(text = "Figure annotation") \%>\%
  add_footer(left = "Program: dm.utils", right = "Page 1 of 1") \%>\%
  grid.draw()

# ... except `clean_slate` and `grid.draw`, the order does not matter
clean_slate(margin = unit(rep(0.05, 4), units = "npc")) \%>\%
  add_note(text = "Figure annotation") \%>\%
  add_header(left = "Topic", right = "CONFIDENTIAL") \%>\%
  add_title("Figure 1") \%>\%
  add_figure(gr = gg) \%>\%
  add_footer(left = "Program: dm.utils", right = "Page 1 of 1") \%>\%
  grid.draw()

# For several lines of annotation, provide a vector
clean_slate(margin = unit(rep(0.05, 4), units = "npc")) \%>\%
  add_figure(gr = gg) \%>\%
  add_header(
    left = c("Topic", "Info / PS9999", "Code"),
    right = c("CONFIDENTIAL", "Draft")
  ) \%>\%
  add_title(text =  c("Figure 1", "title", "Dataset")) \%>\%
  add_note(text = "Figure annotation") \%>\%
  add_footer(left = "Program: dm.utils", right = "Page 1 of 1") \%>\%
  grid.draw()

# By default, notes start at 10\% of width and span over 80 characters
clean_slate(margin = unit(rep(0.05, 4), units = "npc")) \%>\%
  add_figure(gr = gg, width = .5) \%>\%
  add_header(
    left = c("Topic", "Info / PS9999", "Code"),
    right = c("CONFIDENTIAL", "Draft")
  ) \%>\%
  add_title(text = c("Figure 1", "title", "Dataset")) \%>\%
  add_note(
    text = c(
      "\"There are certain queer times and occasions in this strange mixed
      affair we call life when a man takes this whole universe for a vast
      practical joke, though the wit thereof he but dimly discerns, and more
      than suspects that the joke is at nobody's expense but his own.\"",
      "Herman Melville, Moby-Dick"
    )
  ) \%>\%
  add_footer(left = "Program: dm.utils", right = "Page 1 of 1") \%>\%
  grid.draw()

# Check the layout by adding guides
clean_slate(margin = unit(rep(0.05, 4), units = "npc")) \%>\%
  add_figure(gr = gg) \%>\%
  add_header(
    left = c("Topic", "Info / PS9999", "Code"),
    right = c("CONFIDENTIAL", "Draft"),
    guide = TRUE
  ) \%>\%
  add_title(text = c("Figure 1", "title", "Dataset")) \%>\%
  add_note(
    text = c(
      "\"There are certain queer times and occasions in this strange mixed
      affair we call life when a man takes this whole universe for a vast
      practical joke, though the wit thereof he but dimly discerns, and more
      than suspects that the joke is at nobody's expense but his own.\"",
      "Herman Melville, Moby-Dick"
    )
  ) \%>\%
  add_footer(left = "Program: dm.utils", right = "Page 1 of 1") \%>\%
  grid.draw()

}
