% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_diane.R
\name{theme_diane}
\alias{theme_diane}
\title{Theme for ggplot2}
\usage{
theme_diane(...)
}
\arguments{
\item{...}{\cr Arguments passed to [ggplot2::theme()].}
}
\description{
Very sober graphical theme.
}
\examples{

library(ggplot2)
g <- ggplot(mpg, aes(class)) +
  geom_bar()
g
g + theme_diane()

}
